chore: move operator into separate file
